Add styling to the pages
