+++ src/components/navigation/navigation.css
@@ -2,20 +2,28 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
+  padding: 0 12px;
 }
 
 .navbar > h1 {
-
+  font-size: 50px;
 }
 
 .navbar > ul {
   display: flex;
   list-style: none;
-  gap: 20px;
+  gap: 10px;
 }
 
 .navitem {
   color: blue;
   text-decoration: underline;
   cursor: pointer;
+  font-size: 20px;
+  padding: 0 10px;
+  border-right: 3px solid black;
+}
+
+.last {
+  border: none;
 }
