--- src/pages/CalcPage.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import Navigation from '../components/navigation/navigation';
-import Calculator from '../components/calculator/Calculator';
-
-const CalcPage = () => (
-  <div>
-    <Navigation />
-    <Calculator />
-  </div>
-);
-
-export default CalcPage;
