--- src/pages/QuotePage.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import Navigation from '../components/navigation/navigation';
-
-const QuotePage = () => (
-  <div>
-    <Navigation />
-  </div>
-);
-
-export default QuotePage;
