+++ src/pages/homepage/homePage.css
@@ -1,0 +1,3 @@
+.homeContainer {
+    padding: 10px 30px;
+}
