+++ src/pages/quotepage/quotePage.css
@@ -1,0 +1,7 @@
+.quoteContainer {
+    padding: 5rem 10rem;
+}
+
+.quoteText {
+    font-size: 26px;
+}
